+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Parameter `crop_to_image_boundary` to `read_tile()` and `read_encoded_tile()` to allow control of if the tile should be cropped to the image boundary (default, old behavior) or not.
+- Parameter `force_transcoding` to `save()` to control if transcoding should be forced when the image data transfer syntax is the same as the transfer syntax in the `transcoding` parameter.
 
 ### Fixed